Routes, associated functions & files renamed to get_xxxx variable type paramaters removed function test moved
