+++ src/routes/settings/get_network.rs
@@ -16,16 +16,17 @@ use hostname::get_hostname;
 use auth::User;
 
 /// Get Pi-hole local network information
-#[get("/settings/network")]
-pub fn network(env: State<Env>, _auth: User) -> Reply {
+#[get("/settings/get_network")]
+pub fn get_network(env: State<Env>, _auth: User) -> Reply {
     let ipv4_full = read_setup_vars("IPV4_ADDRESS", &env)?.unwrap_or_default();
     let ipv4_address: Vec<&str> = ipv4_full.split("/").collect();
     let ipv6_full = read_setup_vars("IPV6_ADDRESS", &env)?.unwrap_or_default();
     let ipv6_address: Vec<&str> = ipv6_full.split("/").collect();
-    return reply_data(json!({
+
+    reply_data(json!({
         "interface": read_setup_vars("PIHOLE_INTERFACE", &env)?.unwrap_or_default(),
         "ipv4_address": ipv4_address[0],
         "ipv6_address": ipv6_address[0],
         "hostname": get_hostname().unwrap_or("unknown".to_owned())
-    }));
+    }))
 }
