+++ src/routes/settings/mod.rs
@@ -1,11 +1,11 @@
-mod dhcp;
-mod dns;
-mod ftldb;
-mod network;
+mod get_dhcp;
+mod get_dns;
+mod get_ftldb;
+mod get_network;
 mod convert;
 
-pub use self::network::*;
-pub use self::ftldb::*;
-pub use self::dns::*;
-pub use self::dhcp::*;
+pub use self::get_network::*;
+pub use self::get_ftldb::*;
+pub use self::get_dns::*;
+pub use self::get_dhcp::*;
 pub use self::convert::*;
